refactor(html)!: return player element from createPlayer
