+++ packages/html/src/define/audio/player.ts
@@ -1,15 +1,16 @@
 import { audioFeatures } from '@videojs/core/dom';
-
 import { createPlayer } from '../../player/create-player';
 import { ContainerElement } from '../../ui/container/container-element';
-import { UIElement } from '../../ui/ui-element';
 import { safeDefine } from '../safe-define';
 
-const { ProviderMixin } = createPlayer({
+const { PlayerElement, PlayerController: ConfiguredPlayerController } = createPlayer({
   features: audioFeatures,
 });
 
-export class AudioPlayerElement extends ProviderMixin(UIElement) {
+/** Player controller bound to the standard audio player store. */
+export const PlayerController = ConfiguredPlayerController;
+
+export class AudioPlayerElement extends PlayerElement {
   static readonly tagName = 'audio-player';
 }
