+++ packages/html/src/define/live-video/player.ts
@@ -1,15 +1,16 @@
 import { liveVideoFeatures } from '@videojs/core/dom';
-
 import { createPlayer } from '../../player/create-player';
 import { ContainerElement } from '../../ui/container/container-element';
-import { UIElement } from '../../ui/ui-element';
 import { safeDefine } from '../safe-define';
 
-const { ProviderMixin } = createPlayer({
+const { PlayerElement, PlayerController: ConfiguredPlayerController } = createPlayer({
   features: liveVideoFeatures,
 });
 
-export class LiveVideoPlayerElement extends ProviderMixin(UIElement) {
+/** Player controller bound to the live video player store. */
+export const PlayerController = ConfiguredPlayerController;
+
+export class LiveVideoPlayerElement extends PlayerElement {
   static readonly tagName = 'live-video-player';
 }
