+++ packages/html/src/index.ts
@@ -69,7 +69,6 @@ export * from './player/context';
 export * from './player/create-player';
 export * from './player/player-controller';
 export * from './store/media-attach-mixin';
-export * from './store/provider-mixin';
 export * from './store/types';
 export { AirPlayButtonElement } from './ui/airplay-button/airplay-button-element';
 export { AlertDialogBackdropElement } from './ui/alert-dialog/alert-dialog-backdrop-element';
