+++ packages/html/src/player/tests/create-player.test-d.ts
@@ -14,7 +14,7 @@ import {
 import type { Slice } from '@videojs/store';
 import { assertType, describe, it } from 'vite-plus/test';
 
-import { UIElement } from '../../ui/ui-element';
+import type { UIElement } from '../../ui/ui-element';
 import { type CreatePlayerResult, createPlayer } from '../create-player';
 
 describe('createPlayer', () => {
@@ -24,6 +24,12 @@ describe('createPlayer', () => {
     assertType<CreatePlayerResult<VideoPlayerStore>>(result);
     // @ts-expect-error ContainerMixin is no longer part of the HTML API.
     result.ContainerMixin;
+    // @ts-expect-error createPlayer returns a PlayerElement class directly.
+    result.ProviderMixin;
+    // @ts-expect-error Use playerContext.
+    result.context;
+    // @ts-expect-error The player element owns its store.
+    result.create;
   });
 
   it('resolves audio features to AudioPlayerStore', () => {
@@ -31,7 +37,7 @@ describe('createPlayer', () => {
 
     assertType<CreatePlayerResult<AudioPlayerStore>>(result);
 
-    const store = result.create();
+    const store = new result.PlayerElement().store;
 
     assertType<number | undefined>(store.error?.code);
     assertType<string | undefined>(store.error?.message);
@@ -61,31 +67,37 @@ describe('createPlayer', () => {
   it('infers config properties from selected features', () => {
     const withMetadata = createPlayer({ features: [metadataFeature] });
     const withoutMetadata = createPlayer({ features: [features.playback] });
-    const MetadataProvider = withMetadata.ProviderMixin(UIElement);
-    const PlainProvider = withoutMetadata.ProviderMixin(UIElement);
-    const metadataProvider = new MetadataProvider();
-    const plainProvider = new PlainProvider();
+    const metadataPlayer = new withMetadata.PlayerElement();
+    const plainPlayer = new withoutMetadata.PlayerElement();
 
     // The store calls this `title`; on an element that name is the tooltip.
-    assertType<string | null | undefined>(metadataProvider.contentTitle);
+    assertType<string | null | undefined>(metadataPlayer.contentTitle);
 
     // @ts-expect-error metadata properties are absent when the feature is absent.
-    plainProvider.contentTitle;
+    plainPlayer.contentTitle;
+  });
+
+  it('returns a controller already bound to the player context', () => {
+    const { PlayerController } = createPlayer({ features: videoFeatures });
+    const host = null as unknown as UIElement;
+
+    assertType<import('../player-controller').PlayerController<VideoPlayerStore>>(new PlayerController(host));
+    assertType<import('../player-controller').PlayerController<VideoPlayerStore, boolean>>(
+      new PlayerController(host, (state) => state.paused)
+    );
   });
 
   it('exposes orientation lock configuration as a provider property', () => {
     const withOrientationLock = createPlayer({ features: [features.orientationLock] });
     const withoutOrientationLock = createPlayer({ features: [features.playback] });
-    const OrientationProvider = withOrientationLock.ProviderMixin(UIElement);
-    const PlainProvider = withoutOrientationLock.ProviderMixin(UIElement);
-    const orientationProvider = new OrientationProvider();
-    const plainProvider = new PlainProvider();
+    const orientationPlayer = new withOrientationLock.PlayerElement();
+    const plainPlayer = new withoutOrientationLock.PlayerElement();
 
     assertType<CreatePlayerResult<PlayerStore<[typeof features.orientationLock]>>>(withOrientationLock);
-    assertType<ScreenOrientationLockType | null | undefined>(orientationProvider.orientationLockType);
+    assertType<ScreenOrientationLockType | null | undefined>(orientationPlayer.orientationLockType);
 
     // @ts-expect-error orientation lock properties are absent when the feature is absent.
-    plainProvider.orientationLockType;
+    plainPlayer.orientationLockType;
   });
 
   it('resolves extended video features to generic PlayerStore', () => {
@@ -105,7 +117,7 @@ describe('createPlayer', () => {
     assertType<CreatePlayerResult<PlayerStore<[...typeof videoFeatures, typeof analyticsFeature]>>>(result);
 
     // The store has both video and analytics state
-    const store = result.create();
+    const store = new result.PlayerElement().store;
 
     assertType<boolean>(store.paused);
     assertType<number>(store.volume);
@@ -131,7 +143,7 @@ describe('createPlayer', () => {
       features: [...audioFeatures, analyticsFeature],
     });
 
-    const store = result.create();
+    const store = new result.PlayerElement().store;
 
     assertType<boolean>(store.paused);
     assertType<number>(store.volume);
