+++ packages/html/src/player/tests/create-player.test.ts
@@ -33,28 +33,39 @@ describe('createPlayer', () => {
   it('returns expected exports', () => {
     const result = createPlayer({ features: videoFeatures });
 
-    expect(result.context).toBeDefined();
-    expect(result.create).toBeInstanceOf(Function);
+    expect(result.PlayerElement).toBeInstanceOf(Function);
     expect(result.PlayerController).toBeDefined();
-    expect(result.ProviderMixin).toBeInstanceOf(Function);
+    expect(result.playerContext).toBeDefined();
+    expect(result).not.toHaveProperty('ProviderMixin');
+    expect(result).not.toHaveProperty('context');
+    expect(result).not.toHaveProperty('create');
     expect(result).not.toHaveProperty('ContainerMixin');
   });
 
-  it('create() returns a store instance', () => {
-    const { create } = createPlayer({ features: videoFeatures });
-    const store = create();
+  it('PlayerElement owns a store instance', () => {
+    const { PlayerElement } = createPlayer({ features: videoFeatures });
+    const player = document.createElement(defineTestElement(PlayerElement)) as InstanceType<typeof PlayerElement>;
 
-    expect(store.attach).toBeInstanceOf(Function);
-    expect(store.subscribe).toBeInstanceOf(Function);
-    expect(store.destroy).toBeInstanceOf(Function);
+    expect(player.store.attach).toBeInstanceOf(Function);
+    expect(player.store.subscribe).toBeInstanceOf(Function);
+    expect(player.store.destroy).toBeInstanceOf(Function);
   });
 
-  it('ProviderMixin produces a valid custom element class', () => {
-    const { ProviderMixin } = createPlayer({ features: videoFeatures });
-    const ProviderElement = ProviderMixin(UIElement);
+  it('PlayerElement is a valid custom element class', () => {
+    const { PlayerElement } = createPlayer({ features: videoFeatures });
 
-    expect(typeof ProviderElement).toBe('function');
-    expect(ProviderElement.prototype).toBeDefined();
+    expect(typeof PlayerElement).toBe('function');
+    expect(PlayerElement.prototype).toBeDefined();
+  });
+
+  it('preserves author player layout styles', () => {
+    const { PlayerElement } = createPlayer({ features: backgroundFeatures });
+    const player = document.createElement(defineTestElement(PlayerElement));
+
+    player.style.display = 'grid';
+    document.body.append(player);
+
+    expect(player.style.display).toBe('grid');
   });
 
   it('exports a valid ContainerElement class', () => {
@@ -63,7 +74,7 @@ describe('createPlayer', () => {
   });
 
   it('scopes popup coordination to container descendants', async () => {
-    const { ProviderMixin } = createPlayer({ features: videoFeatures });
+    const { PlayerElement } = createPlayer({ features: videoFeatures });
 
     class PopupGroupProbe extends UIElement {
       popupGroup: PopupGroup | undefined;
@@ -79,20 +90,20 @@ describe('createPlayer', () => {
       }
     }
 
-    const providerTag = defineTestElement(ProviderMixin(UIElement));
+    const playerTag = defineTestElement(PlayerElement);
     const containerTag = defineTestElement(ContainerElement);
     const probeTag = defineTestElement(PopupGroupProbe);
-    const provider = document.createElement(providerTag);
+    const player = document.createElement(playerTag);
     const container = document.createElement(containerTag);
     const outsideProbe = document.createElement(probeTag) as PopupGroupProbe;
     const insideProbe = document.createElement(probeTag) as PopupGroupProbe;
 
     container.append(insideProbe);
-    provider.append(outsideProbe, container);
-    document.body.append(provider);
+    player.append(outsideProbe, container);
+    document.body.append(player);
 
     await Promise.all([
-      (provider as UIElement).updateComplete,
+      (player as UIElement).updateComplete,
       (container as UIElement).updateComplete,
       outsideProbe.updateComplete,
       insideProbe.updateComplete,
@@ -103,8 +114,7 @@ describe('createPlayer', () => {
   });
 
   it('keeps container registration identity-safe', async () => {
-    const { ProviderMixin } = createPlayer({ features: backgroundFeatures });
-    const PlayerElement = ProviderMixin(UIElement);
+    const { PlayerElement } = createPlayer({ features: backgroundFeatures });
     const player = document.createElement(defineTestElement(PlayerElement)) as InstanceType<typeof PlayerElement>;
     const first = document.createElement(defineTestElement(class extends ContainerElement {}));
     const second = document.createElement(defineTestElement(class extends ContainerElement {}));
@@ -124,8 +134,7 @@ describe('createPlayer', () => {
   });
 
   it('upgrades parser-created containers under a connected player', async () => {
-    const { ProviderMixin } = createPlayer({ features: backgroundFeatures });
-    const PlayerElement = ProviderMixin(UIElement);
+    const { PlayerElement } = createPlayer({ features: backgroundFeatures });
     const player = document.createElement(defineTestElement(PlayerElement)) as InstanceType<typeof PlayerElement>;
     const containerTag = `test-late-container-${tagCounter++}`;
 
@@ -137,8 +146,7 @@ describe('createPlayer', () => {
   });
 
   it('keeps custom media registration identity-safe', async () => {
-    const { ProviderMixin } = createPlayer({ features: backgroundFeatures });
-    const PlayerElement = ProviderMixin(UIElement);
+    const { PlayerElement } = createPlayer({ features: backgroundFeatures });
     const player = document.createElement(defineTestElement(PlayerElement)) as InstanceType<typeof PlayerElement>;
     const mediaTag = defineTestElement(MediaAttachMixin(HTMLElement));
     const first = document.createElement(mediaTag);
@@ -157,8 +165,7 @@ describe('createPlayer', () => {
   });
 
   it('tracks native media added, removed, and replaced after connection', async () => {
-    const { ProviderMixin } = createPlayer({ features: backgroundFeatures });
-    const PlayerElement = ProviderMixin(UIElement);
+    const { PlayerElement } = createPlayer({ features: backgroundFeatures });
     const player = document.createElement(defineTestElement(PlayerElement)) as InstanceType<typeof PlayerElement>;
     const first = document.createElement('video');
     const second = document.createElement('audio');
@@ -177,8 +184,7 @@ describe('createPlayer', () => {
   });
 
   it('does not retain disconnected context media as a native fallback', async () => {
-    const { ProviderMixin } = createPlayer({ features: backgroundFeatures });
-    const PlayerElement = ProviderMixin(UIElement);
+    const { PlayerElement } = createPlayer({ features: backgroundFeatures });
     const player = document.createElement(defineTestElement(PlayerElement)) as InstanceType<typeof PlayerElement>;
     const background = document.createElement(defineTestElement(BackgroundVideo));
     const video = document.createElement('video');
@@ -198,38 +204,32 @@ describe('createPlayer', () => {
   it('creates audio player with expected exports', () => {
     const result = createPlayer({ features: audioFeatures });
 
-    expect(result.context).toBeDefined();
-    expect(result.create).toBeInstanceOf(Function);
+    expect(result.PlayerElement).toBeInstanceOf(Function);
     expect(result.PlayerController).toBeDefined();
-    expect(result.ProviderMixin).toBeInstanceOf(Function);
-    expect(result).not.toHaveProperty('ContainerMixin');
+    expect(result.playerContext).toBeDefined();
   });
 
   it('creates background player with expected exports', () => {
     const result = createPlayer({ features: backgroundFeatures });
 
-    expect(result.context).toBeDefined();
-    expect(result.create).toBeInstanceOf(Function);
+    expect(result.PlayerElement).toBeInstanceOf(Function);
     expect(result.PlayerController).toBeDefined();
-    expect(result.ProviderMixin).toBeInstanceOf(Function);
-    expect(result).not.toHaveProperty('ContainerMixin');
+    expect(result.playerContext).toBeDefined();
   });
 
   it('maps selected feature inputs to reactive properties and attributes', async () => {
-    const { ProviderMixin } = createPlayer({ features: [metadataFeature] });
-    const ProviderElement = ProviderMixin(UIElement);
+    const { PlayerElement } = createPlayer({ features: [metadataFeature] });
     const tagName = 'test-metadata-provider';
 
-    customElements.define(tagName, ProviderElement);
+    customElements.define(tagName, PlayerElement);
 
-    const player = document.createElement(tagName) as InstanceType<typeof ProviderElement>;
-
+    const player = document.createElement(tagName) as InstanceType<typeof PlayerElement>;
     player.setAttribute('content-title', 'Attribute title');
     document.body.append(player);
 
     expect(player.contentTitle).toBe('Attribute title');
     expect(player.store.title).toBe('Attribute title');
-    expect(ProviderElement.observedAttributes).toContain('content-title');
+    expect(PlayerElement.observedAttributes).toContain('content-title');
 
     player.contentTitle = 'Property title';
 
@@ -250,19 +250,17 @@ describe('createPlayer', () => {
   });
 
   it('leaves the element its own `title`, which means the tooltip', async () => {
-    const { ProviderMixin } = createPlayer({ features: [metadataFeature] });
-    const ProviderElement = ProviderMixin(UIElement);
+    const { PlayerElement } = createPlayer({ features: [metadataFeature] });
     const tagName = 'test-title-provider';
 
-    customElements.define(tagName, ProviderElement);
+    customElements.define(tagName, PlayerElement);
 
-    const player = document.createElement(tagName) as InstanceType<typeof ProviderElement>;
-
+    const player = document.createElement(tagName) as InstanceType<typeof PlayerElement>;
     document.body.append(player);
 
     // Nothing was installed over the native accessor, so this is still the tooltip.
-    expect(Object.getOwnPropertyDescriptor(ProviderElement.prototype, 'title')).toBeUndefined();
-    expect(ProviderElement.observedAttributes).not.toContain('title');
+    expect(Object.getOwnPropertyDescriptor(PlayerElement.prototype, 'title')).toBeUndefined();
+    expect(PlayerElement.observedAttributes).not.toContain('title');
 
     player.title = 'Tooltip';
     await player.updateComplete;
@@ -272,13 +270,12 @@ describe('createPlayer', () => {
   });
 
   it('applies orientation lock configuration through attributes and properties', async () => {
-    const { ProviderMixin } = createPlayer({ features: [features.orientationLock] });
-    const ProviderElement = ProviderMixin(UIElement);
+    const { PlayerElement } = createPlayer({ features: [features.orientationLock] });
     const tagName = 'test-orientation-lock-provider';
 
-    customElements.define(tagName, ProviderElement);
+    customElements.define(tagName, PlayerElement);
 
-    const player = document.createElement(tagName) as InstanceType<typeof ProviderElement>;
+    const player = document.createElement(tagName) as InstanceType<typeof PlayerElement>;
 
     player.setAttribute('orientation-lock-type', 'portrait');
     document.body.append(player);
@@ -299,14 +296,13 @@ describe('createPlayer', () => {
   });
 
   it('applies property configuration that shadowed an accessor before connection', async () => {
-    const { ProviderMixin } = createPlayer({ features: [features.orientationLock] });
-    const ProviderElement = ProviderMixin(UIElement);
-    const tagName = `test-shadowed-config-provider-${tagCounter++}`;
+    const { PlayerElement } = createPlayer({ features: [features.orientationLock] });
+    const tagName = `test-shadowed-config-player-${tagCounter++}`;
 
-    customElements.define(tagName, ProviderElement);
+    customElements.define(tagName, PlayerElement);
 
-    // SAFETY: The tag was registered with ProviderElement immediately above.
-    const player = document.createElement(tagName) as InstanceType<typeof ProviderElement>;
+    // SAFETY: The tag was registered with PlayerElement immediately above.
+    const player = document.createElement(tagName) as InstanceType<typeof PlayerElement>;
 
     Object.defineProperty(player, 'orientationLockType', {
       value: 'portrait',
@@ -325,10 +321,9 @@ describe('createPlayer', () => {
   });
 
   it('leaves config attributes inert when their feature is absent', () => {
-    const { ProviderMixin } = createPlayer({ features: backgroundFeatures });
-    const ProviderElement = ProviderMixin(UIElement);
+    const { PlayerElement } = createPlayer({ features: backgroundFeatures });
 
-    expect(ProviderElement.observedAttributes).not.toContain('content-title');
-    expect(ProviderElement.prototype).not.toHaveProperty('contentTitle');
+    expect(PlayerElement.observedAttributes).not.toContain('content-title');
+    expect(PlayerElement.prototype).not.toHaveProperty('contentTitle');
   });
 });
